Implement triggers for capabilities

+++ backend/services/src/main/java/org/mabartos/persistence/jpa/model/services/device/DeviceServiceImpl.java
@@ -126,7 +126,7 @@ public class DeviceServiceImpl extends CRUDServiceImpl<DeviceModel, DeviceEntity
             device.setRoom(null);
             clearRetainedMessages(deviceID);
             room.removeChild(device);
-            sendDeviceInRoomMessage(roomID, (long) -1);
+            sendDeviceInRoomMessage(roomID, -1L);
             return updateByID(deviceID, device) != null;
         }
         return false;
